refactor: MessageResponse common 패키지로 이동
- 공통 필드를 common 패키지로 이동
