--- backend/src/main/java/dev/chrono/chronochallenge/auth/dto/MessageResponse.java
@@ -1,18 +1,0 @@
-package dev.chrono.chronochallenge.auth.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class MessageResponse {
-
-    private String message;
-
-    private int code;
-
-    @Builder
-    public MessageResponse(String message, int code) {
-        this.message = message;
-        this.code = code;
-    }
-}
